API

GPDiffEq.DerivativeGP — Type

DerivativeGP

The Gaussian Process (GP)

GPDiffEq.GPODE — Type

Defines an ODE model with a GP vector field. The returned model can be evaluated for an initial value as input.

    GPODE(model, tspan, args...; kwargs...)

Arguments:

model: An AbstractGP.FiniteGP

tspan: The time span for the ODE solver

args: Additional positional arguments from DifferentialEquations.jl

kwargs: Additional keyword arguments from DifferentialEquations.jl

AbstractGPs.posterior — Method

posterior(fx::FiniteGP{<:PosteriorGP}, y::AbstractVector{<:Real})

Construct the posterior distribution over fx.f when f is itself a PosteriorGP by updating the Cholesky factorisation of the covariance matrix and avoiding recomputing it from the original covariance matrix. It does this by using update_chol functionality. Other aspects are similar to a regular posterior.

AbstractGPs.posterior — Method

posterior(FiniteGP{<:DerivativeGP}, y::AbstractVector{<:Real})

The posterior of a derivative GP, conditioned on the data y from the output space of the undifferentiated GP. Evaluating this posterior at a point x will return the posterior of the derivative at x, and therefore not return the original data y.

GPDiffEq.differentiate — Method

differentiate(gpp::PosteriorGP)

GPDiffEq.gp_negloglikelihood — Method

Returns two function function. 1) to compute the loglikelihood and 2) rebuild the GP from a new set of parameters.

    gp_negloglikelihood(gp, x, y)

Arguments:

gp: An AbstractGP.FiniteGP

x: input data

y: output data

GPDiffEq.gp_train — Function

Simplified copy of sciml_train

gp_train(loss, θ, opt=ADAGrad(0.5), adtype=Optimization.AutoZygote(), args...;
lower_bounds=nothing, upper_bounds=nothing, maxiters=1000, kwargs...,)